refactor: 홈페이지 및 약관 동의 페이지 UI 수정 (#131)
